Collapse interfaces so it's an interface per endpoint
