--- client/src/main/java/com/lmax/solana4j/client/api/Data.java
@@ -1,10 +1,0 @@
-package com.lmax.solana4j.client.api;
-
-import java.util.List;
-
-public interface Data
-{
-    String getProgramId();
-
-    List<String> getData();
-}
