--- client/src/main/java/com/lmax/solana4j/client/api/InstructionParsed.java
@@ -1,18 +1,0 @@
-package com.lmax.solana4j.client.api;
-
-public interface InstructionParsed
-{
-    InstructionInfoParsed getInfo();
-
-    String getType();
-
-    interface InstructionInfoParsed
-    {
-        String getDestination();
-
-        long getLamports();
-
-        String getSource();
-    }
-
-}
