--- client/src/main/java/com/lmax/solana4j/client/api/SolanaClientError.java
@@ -1,8 +1,0 @@
-package com.lmax.solana4j.client.api;
-
-public interface SolanaClientError
-{
-    long getErrorCode();
-
-    String getErrorMessage();
-}
