--- client/src/main/java/com/lmax/solana4j/client/jsonrpc/DataDTO.java
@@ -1,30 +1,0 @@
-package com.lmax.solana4j.client.jsonrpc;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.lmax.solana4j.client.api.Data;
-
-import java.util.List;
-
-final class DataDTO implements Data
-{
-    private final String programId;
-    private final List<String> data;
-
-    DataDTO(final @JsonProperty("programId") String programId, final @JsonProperty("data") List<String> data)
-    {
-        this.programId = programId;
-        this.data = data;
-    }
-
-    @Override
-    public String getProgramId()
-    {
-        return programId;
-    }
-
-    @Override
-    public List<String> getData()
-    {
-        return data;
-    }
-}
