+++ client/src/main/java/com/lmax/solana4j/client/jsonrpc/SimulateTransactionResponseDTO.java
@@ -2,13 +2,11 @@ package com.lmax.solana4j.client.jsonrpc;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.lmax.solana4j.client.api.AccountInfoData;
+import com.lmax.solana4j.client.api.AccountInfo;
 import com.lmax.solana4j.client.api.Blockhash;
-import com.lmax.solana4j.client.api.Context;
-import com.lmax.solana4j.client.api.Data;
-import com.lmax.solana4j.client.api.InnerInstruction;
 import com.lmax.solana4j.client.api.SimulateTransactionResponse;
 import com.lmax.solana4j.client.api.SolanaRpcResponse;
+import com.lmax.solana4j.client.api.TransactionResponse;
 
 import java.util.List;
 
@@ -51,8 +49,8 @@ final class SimulateTransactionResponseDTO implements SolanaRpcResponse<Simulate
     {
         private final Object err;
         private final List<String> logs;
-        private final AccountInfoData accounts;
-        private final List<InnerInstructionDTO> innerInstructions;
+        private final AccountInfo.AccountInfoData accounts;
+        private final List<TransactionResponseDTO.MetaDTO.InnerInstructionDTO> innerInstructions;
         private final BlockhashDTO.BlockhashValueDTO replacementBlockhash;
         private final DataDTO returnData;
         private final int unitsConsumed;
@@ -61,8 +59,8 @@ final class SimulateTransactionResponseDTO implements SolanaRpcResponse<Simulate
         SimulateTransactionValueDTO(
                 final @JsonProperty("err") Object err,
                 final @JsonProperty("logs") List<String> logs,
-                final @JsonProperty("accounts") AccountInfoData accounts,
-                final @JsonProperty("innerInstructions") List<InnerInstructionDTO> innerInstructions,
+                final @JsonProperty("accounts") AccountInfo.AccountInfoData accounts,
+                final @JsonProperty("innerInstructions") List<TransactionResponseDTO.MetaDTO.InnerInstructionDTO> innerInstructions,
                 final @JsonProperty("replacementBlockhash") BlockhashDTO.BlockhashValueDTO replacementBlockhash,
                 final @JsonProperty("returnData") DataDTO returnData,
                 final @JsonProperty("unitsConsumed") int unitsConsumed)
@@ -89,14 +87,14 @@ final class SimulateTransactionResponseDTO implements SolanaRpcResponse<Simulate
         }
 
         @Override
-        public AccountInfoData getAccounts()
+        public AccountInfo.AccountInfoData getAccounts()
         {
             return accounts;
         }
 
         @Override
         @SuppressWarnings({"unchecked", "rawtypes"})
-        public List<InnerInstruction> getInnerInstructions()
+        public List<TransactionResponse.InnerInstruction> getInnerInstructions()
         {
             return (List) innerInstructions;
         }
@@ -131,6 +129,30 @@ final class SimulateTransactionResponseDTO implements SolanaRpcResponse<Simulate
                     ", returnData=" + returnData +
                     ", unitsConsumed=" + unitsConsumed +
                     '}';
+        }
+    }
+
+    static final class DataDTO implements SimulateTransactionResponse.Data
+    {
+        private final String programId;
+        private final List<String> data;
+
+        DataDTO(final @JsonProperty("programId") String programId, final @JsonProperty("data") List<String> data)
+        {
+            this.programId = programId;
+            this.data = data;
+        }
+
+        @Override
+        public String getProgramId()
+        {
+            return programId;
+        }
+
+        @Override
+        public List<String> getData()
+        {
+            return data;
         }
     }
 }
