+++ client/src/main/java/com/lmax/solana4j/client/jsonrpc/TokenAmountDTO.java
@@ -2,7 +2,6 @@ package com.lmax.solana4j.client.jsonrpc;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.lmax.solana4j.client.api.Context;
 import com.lmax.solana4j.client.api.SolanaRpcResponse;
 import com.lmax.solana4j.client.api.TokenAmount;
